1.loginview needs further Implementation for login direction to search viewmodel.2 finish the basic framework for history function.
