+++ src/user_manage/service/history/service/add_history/AddingHistoryInputData.java
@@ -1,0 +1,4 @@
+package user_manage.service.history.service.add_history;
+
+public class AddingHistoryInputData {
+}
