+++ src/user_manage/service/history/service/add_history/Interface_adapter/AddingHistoryController.java
@@ -1,0 +1,4 @@
+package user_manage.service.history.service.add_history.Interface_adapter;
+
+public class AddingHistoryController {
+}
