+++ src/user_manage/service/history/service/read_history/ReadingHistoryInputBoundary.java
@@ -1,0 +1,6 @@
+package user_manage.service.history.service.read_history;
+
+public interface ReadingHistoryInputBoundary {
+
+    void excute(ReadingHistoryInputData readingHistoryInputData);
+}
