+++ src/user_manage/service/history/service/read_history/ReadingHistoryInputData.java
@@ -1,0 +1,4 @@
+package user_manage.service.history.service.read_history;
+
+public class ReadingHistoryInputData {
+}
