+++ src/user_manage/service/history/service/read_history/interface_adpter/ReadingHistoryPresenter.java
@@ -1,0 +1,4 @@
+package user_manage.service.history.service.read_history.interface_adpter;
+
+public class ReadingHistoryPresenter {
+}
