--- src/user_manage/service/reading_history/ReadingHistoryInputBoundary.java
@@ -1,4 +1,0 @@
-package user_manage.service.reading_history;
-
-public interface ReadingHistoryInputBoundary {
-}
