--- src/user_manage/service/reading_history/ReadingHistoryOutputBoundary.java
@@ -1,4 +1,0 @@
-package user_manage.service.reading_history;
-
-public interface ReadingHistoryOutputBoundary {
-}
