--- src/user_manage/service/reading_history/interactor_adpter/ReadingHistoryState.java
@@ -1,4 +1,0 @@
-package user_manage.service.reading_history.interactor_adpter;
-
-public class ReadingHistoryState {
-}
